+++ dynamics/core.py
